pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.11.0",
    "python-dotenv>=1.1.0",
    "python-ffmpeg>=2.0.12",
]
